Added fee currency option
